+++ src/components/App/App.scss
@@ -1,18 +1,4 @@
-$textColor: #3c3c3c;
-$hoverBtnColor: #f9c910;
-$sectionsBgColor: #ebebeb;
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  color: $textColor;
-  font-weight: 300;
-}
-
-.App {
-  font-family: 'Open Sans', sans-serif;
-}
+@import '../../styles/variables';
 
 .container {
   width: 85%;
@@ -38,8 +24,9 @@ $sectionsBgColor: #ebebeb;
 .section-steps,
 .section-AboutUs,
 .section-info,
-.form{
-  width: 1900px;
+.form {
+  //uzyc max-width
+  max-width: 1900px;
   margin: 0 auto;
 }
 
@@ -50,7 +37,7 @@ $sectionsBgColor: #ebebeb;
 
   &:after {
     margin: 20px auto;
-    content: url("../img/decoration.png");
+    content: url("../../img/decoration.png");
     width: 253px;
     height: 50px;
     display: block;
@@ -70,13 +57,10 @@ $sectionsBgColor: #ebebeb;
 }
 
 
-
-
-
-
 .header {
-  height: 964px;
-  background: url("../img/header_start_img.jpg") no-repeat;
+  height: auto;
+  background: url("../../img/header_start_img.jpg") no-repeat;
+  background-size: auto;
 
   .container {
     display: flex;
@@ -139,12 +123,13 @@ $sectionsBgColor: #ebebeb;
 
   .btn {
     font-size: 36px;
+    background: rgba(255, 255, 255, 0.5);
   }
 }
 
 .section {
   height: 382px;
-  background: url("../img/Bg-yellow.png");
+  background: url("../../img/Bg-yellow.png");
   background-position: center;
 
   .container {
@@ -174,90 +159,6 @@ $sectionsBgColor: #ebebeb;
   }
 }
 
-.section-steps {
-
-  .container {
-    height: 700px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    h1 {
-      margin-top: 80px;
-    }
-
-    .steps-row {
-      width: 1180px;
-      height: 325px;
-      background-color: $sectionsBgColor;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-
-      .col-1-4 {
-        text-align: center;
-        margin: 0 70px;
-
-        div {
-          margin: 0 auto;
-          width: 80px;
-          height: 70px;
-        }
-
-        #img28 {
-          background: url("../img/icons/28.png") no-repeat center;
-        }
-
-        #img016 {
-          background: url("../img/icons/016.png") no-repeat center;
-        }
-
-        #img39 {
-          background: url("../img/icons/39.png") no-repeat center;
-        }
-
-        #img58 {
-          background: url("../img/icons/58.png") no-repeat center;
-        }
-
-        h3 {
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          width: auto;
-          height: 52px;
-          font-size: 18px;
-          font-weight: 400;
-          letter-spacing: -0.36px;
-        }
-
-        hr {
-          margin: 10px auto;
-          width: 88px;
-          border: 1px solid $textColor;
-        }
-
-        p {
-          width: 170px;
-          height: 50px;
-          color: #3c3c3c;
-          font-size: 16px;
-          font-weight: 400;
-          line-height: 18px;
-
-        }
-      }
-    }
-
-    .btn {
-      margin: 50px 0;
-      font-size: 36px;
-      font-weight: 300;
-      text-transform: uppercase;
-    }
-  }
-}
-
 .section-AboutUs {
   background: $sectionsBgColor;
   height: 967px;
@@ -268,26 +169,34 @@ $sectionsBgColor: #ebebeb;
     width: 100%;
 
     .section-AboutUs-box {
-      height: 600px;
-      width: 600px;
-      margin-top: 250px;
-      margin-left: 207px;
+      display: flex;
+      flex-direction: column;
+      height: 80%;
+      width: 80%;
       text-align: center;
+      margin: 150px auto;
+      padding: 0 50px;
 
       .section-AboutUs-text {
+        padding: 30px 0;
         font-size: 30px;
         font-weight: 400;
         line-height: 54px;
       }
 
       #imgSignature {
-        margin-top: 50px;
-        margin-left: 250px;
+        display: flex;
+        margin-left: 50%;
+        object-fit: contain;
+
       }
     }
 
     #imgAbout {
+      object-fit: contain;
       margin: 55px 0;
+      width: 50%;
+      height: auto;
     }
   }
 }
@@ -297,56 +206,81 @@ $sectionsBgColor: #ebebeb;
     display: flex;
     flex-direction: column;
     align-items: center;
+
     .section-info-header {
       text-align: center;
       width: 600px;
       margin-top: 100px;
+
       nav {
         display: flex;
         justify-content: space-between;
+
         button {
           height: 78px;
           font-size: 24px;
           font-weight: 400;
           border: none;
-          &:focus{
+
+          &:focus {
             border: 1px solid;
           }
         }
       }
+
       p {
         font-size: 22px;
         font-weight: 400;
         margin: 50px 0;
       }
     }
-    .info-section{
+
+    .info-section {
       display: flex;
       justify-content: space-between;
       width: 1266px;
+
       hr {
         border: 1px solid $textColor;
         width: 496px;
         margin: 33px 0;
       }
+
       span {
         display: block;
         font-size: 29px;
         font-family: 'Merriweather', serif;
         font-weight: 300;
       }
+
       p {
         font-family: "Merriweather - Light Italic", serif;
         font-size: 18px;
         font-weight: 400;
+        position: relative;
+        margin-bottom: 50px;
+
+        &:after {
+          content: "";
+          display: block;
+          position: absolute;
+          left: 0;
+          bottom: -20px;
+          height: 1px;
+          width: 100%;
+          background: black;
+        }
       }
+
       .info-left {
         margin: auto 0;
       }
+
       .info-right {
         margin: 0;
         text-align: right;
-        p{
+
+        p {
           display: flex;
           align-items: flex-end;
           justify-content: flex-end;
@@ -354,11 +288,13 @@ $sectionsBgColor: #ebebeb;
         }
       }
     }
-    ul{
+
+    ul {
       display: flex;
       font-size: 24px;
       font-weight: 400;
-      li{
+
+      li {
         list-style: none;
         height: 42px;
         width: 35px;
@@ -368,51 +304,97 @@ $sectionsBgColor: #ebebeb;
 }
 
 .form {
-  height: 958px;
-  background: #eeefef url("../img/footer_bg_img.jpg") no-repeat;
-  .container{
+  height: 800px;
+  background: #eeefef url("../../img/footer_bg_img.jpg") no-repeat;
+  background-size: 80% 100%;
+
+  .container {
     display: flex;
-    justify-content: flex-end;
+    align-items: center;
+    flex-direction: column;
+    justify-content: center;
+    height: 100%;
+    width: 80%;
 
+    h1 {
+      margin: 100px;
+    }
 
-    h4{
-      margin-top: 200px;
+    h4 {
       text-transform: uppercase;
     }
+
     form {
-      margin: auto 0;
+      height: 100%;
+      align-self: flex-end;
     }
-    .form-row{
+
+    .form-row {
       display: flex;
       justify-content: space-between;
-      width: 570px;
     }
+
     #email,
     #name {
       width: 45%;
+      &::placeholder {
+        color: $textColor;
+      }
     }
+
     .form-control {
       padding: 10px;
       background: none;
       font-size: 18px;
-      outline:none;
+      outline: none;
       margin: 20px 0;
       border: none;
       border-bottom: 1px solid;
     }
-    #message{
+
+    #message {
       width: 100%;
+      &::placeholder {
+        color: $textColor;
+      }
     }
   }
-  .form-btn{
+
+  .form-btn {
     display: flex;
     justify-content: center;
   }
+
   button {
     display: flex;
     justify-content: center;
     margin: 20px 0;
     width: 100px;
     height: 50px;
+  }
+
+  footer {
+    width: 80%;
+    height: 50px;
+    color: #3c3c3c;
+    font-family: Merriweather, serif;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+
+    h2 {
+      text-align: center;
+      font-weight: 300;
+      width: 90%;
+    }
+    img {
+      margin: 0 5px;
+      border: 1px solid black;
+      padding: 10px;
+      height: 40px;
+      width: 40px;
+      object-fit: contain;
+    }
   }
 }
