+++ src/components/Header/Header.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Nav from './nav.js'
+import Nav from '../Nav/Nav'
 
 export default class Header extends React.Component {
     render() {
