+++ src/components/Section/Section.js
@@ -3,12 +3,12 @@ import React from 'react'
 export default class Section extends React.Component {
     render() {
         return (
-            <div className={'section'}>
-                <div className={'container'}>
-                    <div className={'col-1-3'}>
-                        <p className={'numbers'}>10</p>
-                        <h2 className={'numbers-name'}>Oddanych worków</h2>
-                        <div className={'numbers-text'}>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aspernatur commodi culpa deserunt do</div>
+            <div className='section'>
+                <div className='container'>
+                    <div className='col-1-3'>
+                        <p className='numbers'>10</p>
+                        <h2 className='numbers-name'>Oddanych worków</h2>
+                        <div className='numbers-text'>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aspernatur commodi culpa deserunt do</div>
                     </div>
                     <div className={'col-1-3'}>
                         <p className={'numbers'}>5</p>
