+++ src/styles/variables.scss
@@ -1,0 +1,3 @@
+$textColor: #3c3c3c;
+$hoverBtnColor: #f9c910;
+$sectionsBgColor: #ebebeb;
